fix: make symlinke in order to link to `libta-lib.so`
For some reason, in the recent build pip couldn't install TA-Lib
with error '/usr/bin/ld: cannot find -lta-lib: No such file or directory'.

The TA-Lib library is located at '/usr/bin/libta_lib.so' while TA-Lib
pip package install expects '.../libta-lib.so'.

I guess that newer version of TA-Lib builds with different name, thus
pip install redirected to the new name to reflect the changes. But for
the moment I tried not to move to be pre-built version of TA-Lib library
as I'm not sure if it's still support armv7 (32 bit Raspberry Pi), so I
did a workaround by symlink the built library with the new name.